Rename Exceptions and get functions
